% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{add_sum_contaminants}
\alias{add_sum_contaminants}
\title{Title}
\usage{
add_sum_contaminants(
  .cols,
  .na_values = c(-9, -99, -99.99),
  .na_numeric = -99.99,
  na.rm = TRUE
)
}
\arguments{
\item{na.rm}{}
}
\description{
Title
}
