% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_uncertainty.R
\name{calculate_uncertainty_2}
\alias{calculate_uncertainty_2}
\title{Title}
\usage{
calculate_uncertainty_2(data, column, cut, const1, const2)
}
\arguments{
\item{const2}{}
}
\description{
Title
}
