% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egg_variables.R
\name{egg_vol}
\alias{egg_vol}
\title{Title}
\usage{
egg_vol(
  egg_length,
  egg_width,
  vol_type = c("outer", "inner"),
  shell_thickness = NULL,
  a = 0.0373,
  b = 35.3
)
}
\arguments{
\item{b}{}
}
\description{
Title
}
