% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_egg_variables.R
\name{egg_volume}
\alias{egg_volume}
\title{Title}
\usage{
egg_volume(
  egg_length,
  egg_width,
  shell_thickness,
  unit = NULL,
  type = NULL,
  replace_na_shell_thickness = TRUE,
  a = 0.0373,
  b = 35.3
)
}
\arguments{
\item{b}{}
}
\description{
Title
}
