% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{eggs_from_excel}
\alias{eggs_from_excel}
\title{Title}
\usage{
eggs_from_excel(
  path = "eggs.xlsx",
  sheet = "Eggs",
  length_unit = c("cm", "mm")
)
}
\arguments{
\item{path}{}

\item{sheet}{}

\item{length_unit}{}
}
\description{
Title
}
