% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_range_excel.R
\name{get_data_range}
\alias{get_data_range}
\title{Title}
\usage{
get_data_range(
  path,
  sheet,
  start_chr,
  start_row = NULL,
  end_chr = NULL,
  start_col = 1,
  stop_col = NULL,
  as.range = TRUE,
  n = 1,
  search_column = character
)
}
\arguments{
\item{search_column}{}
}
\description{
Title
}
