% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_lists.R
\name{get_samples_ov_xl}
\alias{get_samples_ov_xl}
\title{Title}
\usage{
get_samples_ov_xl(
  path = "esb/excel_catalogs/tmp_OV_katalog_havsorn.xlsx",
  sheet = "samples"
)
}
\arguments{
\item{sheet}{}
}
\description{
Title
}
