% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_lists.R
\name{get_samples_species}
\alias{get_samples_species}
\title{Title}
\usage{
get_samples_species(
  codes_path = system.file("extdata", "codelist_wtse.xlsx", package = "MoCiS2.wtse")
)
}
\arguments{
\item{codes_path}{}
}
\description{
Title
}
