% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_itm_aces_templates.R
\name{import_new_template}
\alias{import_new_template}
\title{Title}
\usage{
import_new_template(
  path,
  sheet,
  range,
  col_names = TRUE,
  cols_first_5 = c("accnr", "labcode", "species", "sampling_site", "sample_tissue")
)
}
\arguments{
\item{cols_first_5}{}
}
\description{
Title
}
