% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_itm_aces_templates.R
\name{import_new_template_data}
\alias{import_new_template_data}
\title{Title}
\usage{
import_new_template_data(path, sheet = c("results", "uncertainty"), range)
}
\arguments{
\item{range}{}
}
\description{
Title
}
