% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_itm_aces_templates.R
\name{read_excel_search}
\alias{read_excel_search}
\title{Title}
\usage{
read_excel_search(
  path,
  sheet,
  search_str,
  i = 1,
  n_rows = NA,
  col_names = TRUE
)
}
\arguments{
\item{col_names}{}
}
\description{
Title
}
