% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_itm_aces_templates.R
\name{read_lab_file_itm_header}
\alias{read_lab_file_itm_header}
\title{Title}
\usage{
read_lab_file_itm_header(
  path,
  sheet = "summery",
  range_1 = "A1:D7",
  range_2 = "G1:J4",
  range_3 = "K1:O5"
)
}
\arguments{
\item{range_3}{}
}
\description{
Title
}
