% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variables.R
\name{rename_contaminants}
\alias{rename_contaminants}
\title{Title}
\usage{
rename_contaminants(
  .data,
  .pattern = c(`\\\\s+` = "", `,` = "", `\\\\*` = "", `-` = "", FPRC = "fat_percentage",
    `K$` = "", `\\\\(` = "", `\\\\)` = "", `\\\\\%` = "", `\\\\/` = "_", `^header$` =
    "accnr", `LABKODICB$` = "labcode", `LABKOD$` = "labcode", `^site$` = "samplingsite"),
  .cols = tidyselect::everything()
)
}
\arguments{
\item{.cols}{}
}
\description{
Title
}
