% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{wtse_eggs_from_esbase}
\alias{wtse_eggs_from_esbase}
\title{Title}
\usage{
wtse_eggs_from_esbase(
  path = "esb/esbase/esbase_dump.db",
  extended_types = TRUE,
  query = "SELECT * FROM v_egg WHERE species_id = 597",
  length_unit = c("mm", "cm")
)
}
\arguments{
\item{length_unit}{}
}
\description{
Title
}
