% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{wtse_eggs_from_excel}
\alias{wtse_eggs_from_excel}
\title{Title}
\usage{
wtse_eggs_from_excel(
  path = "eggs.xlsx",
  sheet = "eggs",
  skip = 1,
  rename = TRUE,
  collector_org = "PROVTAG_ORG",
  accnr_only = TRUE
)
}
\arguments{
\item{accnr_only}{}
}
\description{
Title
}
