% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{wtse_eggs_from_ov}
\alias{wtse_eggs_from_ov}
\title{Title}
\usage{
wtse_eggs_from_ov(
  path = "esb/excel_catalogs/OV-Havsörn.xlsx",
  sheet = "OV Havsörn (edit)",
  output = c("samples", "specimens"),
  remove_empty_cols = TRUE
)
}
\arguments{
\item{remove_empty_cols}{}
}
\description{
Title
}
