% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{wtse_sites_eggs}
\alias{wtse_sites_eggs}
\title{Title}
\usage{
wtse_sites_eggs(path = "eggs.xlsx", sheet = "Eggs", include_labcode = FALSE)
}
\arguments{
\item{include_labcode}{}
}
\description{
Title
}
