% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodata_and_provmetadata.R
\name{wtse_sites_provmetadata}
\alias{wtse_sites_provmetadata}
\title{Title}
\usage{
wtse_sites_provmetadata(path = "eggs.xlsx", dsn = "Havsorn_Data")
}
\arguments{
\item{path}{}
}
\description{
Title
}
