% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_data_below_loq.R
\name{fix_below_loq}
\alias{fix_below_loq}
\title{Title}
\usage{
fix_below_loq(
  .x,
  .bound = c("medium", "lower", "upper"),
  .replace = TRUE,
  .replace_values = c(-99.99, -9, -99, 0, NA, NaN, Inf),
  .replace_with = NA
)
}
\arguments{
\item{.replace_na}{}
}
\description{
Title
}
