% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_data_loq_accnr.R
\name{format_accnr}
\alias{format_accnr}
\title{Title}
\usage{
format_accnr(
  x,
  regex = "(?<catalogue>[A-Za-z]{1})(?<accyear>[0-9]{2,4})\\\\/(?<nr>[0-9]{3,5})",
  exclude_non_matches = FALSE
)
}
\arguments{
\item{exclude_non_matches}{}
}
\description{
Title
}
