% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples_lists.R
\name{wtse_eggs_1996}
\alias{wtse_eggs_1996}
\title{Title}
\usage{
wtse_eggs_1996(
  path = "eggs-1996.xlsx",
  sheet = "data",
  range = "A3:AG501",
  sheet_names = "column_names"
)
}
\arguments{
\item{sheet_names}{}
}
\description{
Title
}
